SCT-1091 - Refactoring / moving stuff around
