+++ components/WarningNotes/WarningNoteRecap/WarningNoteRecap.tsx
