+++ components/WarningNotes/WarningNotes.module.scss
