+++ components/WarningNotes/WarningNotes.tsx
