+++ pages/people/[id]/index.tsx
@@ -11,7 +11,7 @@ import { useUnfinishedSubmissions } from 'utils/api/submissions';
 import { canManageCases } from 'lib/permissions';
 import { isAuthorised } from 'utils/auth';
 import { ConditionalFeature } from '../../../lib/feature-flags/feature-flags';
-import CaseStatusView from 'components/pages/people/case-status/view/case-status';
+import CaseStatusView from 'components/CaseStatus/CaseStatusView';
 import Cases from 'components/Cases/Cases';
 
 interface Props {
